Fixing porting issues

+++ aten/src/ATen/native/UpSampling.h
@@ -2,11 +2,14 @@
 
 #include <ATen/ATen.h>
 
-namespace at { namespace native {
+namespace at {
+namespace native {
 
-template<typename T>
+template <typename T>
 static inline T linear_upsampling_compute_scale(
-                          int inputSize, int outputSize, bool align_corners) {
+    int inputSize,
+    int outputSize,
+    bool align_corners) {
   /* We view each pixel as an area, idx + 0.5 as its center index.
    * Here is an example formula in 1D case.
    * if align_corners: center of two corner pixel areas are preserved,
@@ -19,16 +22,18 @@ static inline T linear_upsampling_compute_scale(
    *     src_idx + 0.5 = scale * (dst_index + 0.5)
    */
   if (outputSize > 1) {
-    return align_corners ? (T) (inputSize - 1) / (outputSize - 1)
-                         : (T) inputSize / outputSize;
+    return align_corners ? (T)(inputSize - 1) / (outputSize - 1)
+                         : (T)inputSize / outputSize;
   } else {
     return T(0);
   }
 }
 
-template<typename T>
+template <typename T>
 static inline T linear_upsampling_compute_source_index(
-                          T scale, int dst_index, bool align_corners) {
+    T scale,
+    int dst_index,
+    bool align_corners) {
   if (align_corners) {
     return scale * dst_index;
   } else {
@@ -38,44 +43,51 @@ static inline T linear_upsampling_compute_source_index(
 }
 
 static inline int nearest_neighbor_compute_source_index(
-		const float scale, int dst_index, int inputSize) {
-  const int src_index = std::min(floorf(dst_index * scale), inputSize - 1);
+    const float scale,
+    int dst_index,
+    int inputSize) {
+  const int src_index = std::min(static_cast<int>(floorf(dst_index * scale)), inputSize - 1);
   return src_index;
 }
 
-template<typename T>
-static T upsampling_get_value_bounded(T* data, int width, int height, int x, int y) {
+template <typename T>
+static T upsampling_get_value_bounded(
+    T* data,
+    int width,
+    int height,
+    int x,
+    int y) {
   int access_x = std::max(std::min(x, width - 1), 0);
   int access_y = std::max(std::min(y, height - 1), 0);
   return data[access_y * width + access_x];
 }
 
-template<typename T>
+template <typename T>
 static void upsampling_increment_value_bounded(
-  T* data,
-  int width,
-  int height,
-  int x,
-  int y,
-  T value
-) {
+    T* data,
+    int width,
+    int height,
+    int x,
+    int y,
+    T value) {
   int access_x = std::max(std::min(x, width - 1), 0);
   int access_y = std::max(std::min(y, height - 1), 0);
   data[access_y * width + access_x] += value;
 }
 
-// Based on https://en.wikipedia.org/wiki/Bicubic_interpolation#Bicubic_convolution_algorithm
-template<typename T>
+// Based on
+// https://en.wikipedia.org/wiki/Bicubic_interpolation#Bicubic_convolution_algorithm
+template <typename T>
 static inline T cubic_convolution1(T x, T A) {
   return ((A + 2) * x - (A + 3)) * x * x + 1;
 }
 
-template<typename T>
+template <typename T>
 static inline T cubic_convolution2(T x, T A) {
   return ((A * x - 5 * A) * x + 8 * A) * x - 4 * A;
 }
 
-template<typename T>
+template <typename T>
 static inline void get_cubic_upsampling_coefficients(T coeffs[4], T t) {
   T A = -0.75;
 
@@ -89,21 +101,13 @@ static inline void get_cubic_upsampling_coefficients(T coeffs[4], T t) {
   coeffs[3] = cubic_convolution2<T>(x2 + 1.0, A);
 }
 
-template<typename T>
-static inline T cubic_interp1d(
-  T x0,
-  T x1,
-  T x2,
-  T x3,
-  T t
-) {
+template <typename T>
+static inline T cubic_interp1d(T x0, T x1, T x2, T x3, T t) {
   T coeffs[4];
   get_cubic_upsampling_coefficients<T>(coeffs, t);
 
-  return x0 * coeffs[0]
-    + x1 * coeffs[1]
-    + x2 * coeffs[2]
-    + x3 * coeffs[3];
+  return x0 * coeffs[0] + x1 * coeffs[1] + x2 * coeffs[2] + x3 * coeffs[3];
 }
 
-}}
+} // namespace native
+} // namespace at
